Extract utils for chain
